+++ fortran/src/global.f90
@@ -2,55 +2,55 @@ module global
       implicit none      
 
       !Time Parameters
-      real,parameter :: tmax = 0.0      !Max run time in Seconds
-      real,parameter :: cour = 0.85     !fraction of courant number
+      double precision,parameter :: tmax = 0.0      !Max run time in Seconds
+      double precision,parameter :: cour = 0.85     !fraction of courant number
       
-      real,parameter :: m = 2.0         !azimuthal (wavenumber)
-      real,parameter :: drive0 = 10.0e-9!amplitude of driver (T)
-      real,parameter :: freq = 0.01     !frequency of driver in (Hz)
+      double precision,parameter :: m = 2.0         !azimuthal (wavenumber)
+      double precision,parameter :: drive0 = 10.0e-9!amplitude of driver (T)
+      double precision,parameter :: freq = 0.01     !frequency of driver in (Hz)
 
       !Directories
       character(*),parameter :: data_dir = "./data"     !data directory
       
       !Grid Parameters
-      real,parameter :: Lval = 6.0      !conoical L value
-      real,parameter :: LMin = 2.0      !min L
-      real,parameter :: LMax = 10.0     !max L
+      double precision,parameter :: Lval = 6.0      !conoical L value
+      double precision,parameter :: LMin = 2.0      !min L
+      double precision,parameter :: LMax = 10.0     !max L
       integer,parameter :: num_u1 = 151    !number of field lines
-      real,parameter :: Re = 6378.388e3 !Re in metres
-      real,parameter :: z0 = 80.0e3     !height of ionospheric thin sheet current (in) m
-      !real,parameter :: ds0 = 5.0e3    !grid spacing (along cononical field line) at z0 (in m)
-      real,parameter :: dsn = 500.0e3   !grid spacing (along cononical field line) at rtop (in m)
-      real,parameter :: modefrac = 0.2  !fraction (of gridpoints in u1) to use as basis function in atmospheric expansions
-      real,parameter :: d32 = 2.0 !"spacing" between the k+1 and k-1 points in the u3 direction
+      double precision,parameter :: Re = 6378.388e3 !Re in metres
+      double precision,parameter :: z0 = 80.0e3     !height of ionospheric thin sheet current (in) m
+      !double precision,parameter :: ds0 = 5.0e3    !grid spacing (along cononical field line) at z0 (in m)
+      double precision,parameter :: dsn = 500.0e3   !grid spacing (along cononical field line) at rtop (in m)
+      double precision,parameter :: modefrac = 0.2  !fraction (of gridpoints in u1) to use as basis function in atmospheric expansions
+      double precision,parameter :: d32 = 2.0 !"spacing" between the k+1 and k-1 points in the u3 direction
 
       !constants
-      real,parameter :: pi = 3.141592
+      double precision,parameter :: pi = 3.141592
       !Scaled physical constants for unit length in Re
       
-      real,parameter :: c_u = 2.997e8            !speed of light in vacuum
-      real,parameter :: u0_u = 4.0*pi*1.0e-7     !magnetic permeability (SI units)
-      real,parameter :: e0_u = 1.0/(u0_u*c_u**2) !dielectric constant for a vacuum in (si) units
-      real,parameter :: c2 = (c_u/Re)**2         !speed of light squared in re^2/c^2
-      real,parameter :: mhocgs = 1.0/(4.0*pi*e0_u) !conversion factor needed with bobs ionospheric file 
-      real,parameter :: re2 = re**2              !Re^2
+      double precision,parameter :: c_u = 2.997e8            !speed of light in vacuum
+      double precision,parameter :: u0_u = 4.0*pi*1.0e-7     !magnetic permeability (SI units)
+      double precision,parameter :: e0_u = 1.0/(u0_u*c_u**2) !dielectric constant for a vacuum in (si) units
+      double precision,parameter :: c2 = (c_u/Re)**2         !speed of light squared in re^2/c^2
+      double precision,parameter :: mhocgs = 1.0/(4.0*pi*e0_u) !conversion factor needed with bobs ionospheric file 
+      double precision,parameter :: re2 = re**2              !Re^2
       complex,parameter :: im = (0,m)
 
       !Global Variables
       integer :: num_u3,num_u3_half
-      real :: del_u1,dt
+      double precision :: del_u1,dt
 
       !Global Arrays
-      real,dimension(:,:),allocatable :: x_arr,y_arr,z_arr,r_arr
-      real,dimension(:),allocatable :: hratm_N,hthatm_N,hphiatm_N,hthg_N,hphig_N
-      real,dimension(:),allocatable :: hratm_S,hthatm_S,hphiatm_S,hthg_S,hphig_S
-      real,dimension(:,:),allocatable :: sigp_arr,sigh_arr,sig0_arr,eta_arr
+      double precision,dimension(:,:),allocatable :: x_arr,y_arr,z_arr,r_arr
+      double precision,dimension(:),allocatable :: hratm_N,hthatm_N,hphiatm_N,hthg_N,hphig_N
+      double precision,dimension(:),allocatable :: hratm_S,hthatm_S,hphiatm_S,hthg_S,hphig_S
+      double precision,dimension(:,:),allocatable :: sigp_arr,sigh_arr,sig0_arr,eta_arr
 
 
-      real,dimension(:,:),allocatable :: costh,va2_arr,eps_arr
+      double precision,dimension(:,:),allocatable :: costh,va2_arr,eps_arr
       !Grid Arrays
-      real,dimension(:,:),allocatable :: h3,h1,h2,bsqrt,g22,g11,g13,g21,g12,jac,g33
-      real,dimension(:,:),allocatable :: gsup11,gsup13
-      real,dimension(:),allocatable :: u1
+      double precision,dimension(:,:),allocatable :: h3,h1,h2,bsqrt,g22,g11,g13,g21,g12,jac,g33
+      double precision,dimension(:,:),allocatable :: gsup11,gsup13
+      double precision,dimension(:),allocatable :: u1
 
 end module
